Align source and test packages. Isolate inherited core component tests package so they could be excluded from build.
