+++ vaadin-testbench-it/vaadin-testbench-it-test/src/main/java/com/vaadin/testbench/tests/ui/PerformanceView.java
@@ -1,4 +1,4 @@
-package com.vaadin.testbench.tests.uitest;
+package com.vaadin.testbench.tests.ui;
 
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.NativeButton;
