+++ vaadin-testbench-it/vaadin-testbench-it-test/src/main/java/com/vaadin/testbench/tests/ui/element/PolymerTemplateView.java
@@ -1,4 +1,4 @@
-package com.vaadin.testbench.tests.uitest;
+package com.vaadin.testbench.tests.ui.element;
 
 import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.dependency.HtmlImport;
