--- vaadin-testbench-it/vaadin-testbench-it-test/src/main/java/com/vaadin/testbench/tests/uitest/MyComponentWithClasses.java
@@ -1,13 +1,0 @@
-package com.vaadin.testbench.tests.uitest;
-
-import com.vaadin.flow.component.html.Div;
-
-public class MyComponentWithClasses extends Div {
-
-    public MyComponentWithClasses() {
-        addClassName("my-component-first");
-        addClassName("my-component-with-classes");
-        addClassName("my-component-last");
-        setText(getClass().getSimpleName());
-    }
-}
