--- vaadin-testbench-it/vaadin-testbench-it-test/src/main/java/com/vaadin/testbench/tests/uitest/MyComponentWithId.java
@@ -1,11 +1,0 @@
-package com.vaadin.testbench.tests.uitest;
-
-import com.vaadin.flow.component.html.Div;
-
-public class MyComponentWithId extends Div {
-
-    public MyComponentWithId() {
-        setId("my-component-with-id");
-        setText(getClass().getSimpleName());
-    }
-}
