+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/component/checkbox/CheckboxIT.java
@@ -3,7 +3,7 @@ package junit.com.vaadin.testbench.tests.component.checkbox;
 import com.vaadin.flow.component.checkbox.testbench.CheckboxElement;
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
 import junit.com.vaadin.testbench.tests.component.common.AbstractIT;
-import junit.com.vaadin.testbench.tests.uitest.GenericTestPageObject;
+import junit.com.vaadin.testbench.tests.ui.GenericTestPageObject;
 import org.junit.jupiter.api.Assertions;
 
 import static com.vaadin.flow.component.checkbox.testbench.test.CheckboxView.CHECKED;
