+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/ui/SVGIT.java
@@ -1,4 +1,4 @@
-package junit.com.vaadin.testbench.tests.uitest;
+package junit.com.vaadin.testbench.tests.ui;
 
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
 import com.vaadin.testbench.addons.webdriver.SkipBrowsers;
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Assertions;
 import org.openqa.selenium.By;
 
 import static com.vaadin.testbench.addons.webdriver.BrowserTypes.SAFARI;
-import static com.vaadin.testbench.tests.uitest.SVGView.ROUTE;
+import static com.vaadin.testbench.tests.ui.SVGView.ROUTE;
 
 @VaadinTest
 class SVGIT {
