+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/ui/demo/MainViewIT.java
@@ -1,4 +1,4 @@
-package junit.com.vaadin.testbench.tests.demo;
+package junit.com.vaadin.testbench.tests.ui.demo;
 
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
 import org.junit.jupiter.api.DisplayName;
