+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/ui/element/BasicElementIT.java
@@ -1,9 +1,9 @@
-package junit.com.vaadin.testbench.tests.uitest.elements;
+package junit.com.vaadin.testbench.tests.ui.element;
 
 import com.vaadin.testbench.TestBenchElement;
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
-import com.vaadin.testbench.tests.uitest.ElementQueryView;
-import junit.com.vaadin.testbench.tests.uitest.GenericTestPageObject;
+import com.vaadin.testbench.tests.ui.element.ElementQueryView;
+import junit.com.vaadin.testbench.tests.ui.GenericTestPageObject;
 import org.junit.jupiter.api.Assertions;
 import org.openqa.selenium.TimeoutException;
