+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/ui/element/ElementScreenCompareIT.java
@@ -1,15 +1,15 @@
-package junit.com.vaadin.testbench.tests.uitest.elements;
+package junit.com.vaadin.testbench.tests.ui.element;
 
 import com.vaadin.testbench.TestBenchElement;
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
-import junit.com.vaadin.testbench.tests.uitest.GenericTestPageObject;
+import junit.com.vaadin.testbench.tests.ui.GenericTestPageObject;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.openqa.selenium.By;
 import org.openqa.selenium.OutputType;
 
-import static com.vaadin.testbench.tests.uitest.ElementQueryView.ROUTE;
+import static com.vaadin.testbench.tests.ui.element.ElementQueryView.ROUTE;
 
 @VaadinTest
 @Disabled("Throws an exception related to file I/O. Should be fixed.")
