+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/ui/element/ExecuteJavaScriptIT.java
@@ -1,11 +1,11 @@
-package junit.com.vaadin.testbench.tests.uitest.elements;
+package junit.com.vaadin.testbench.tests.ui.element;
 
 import com.vaadin.testbench.TestBenchElement;
 import com.vaadin.testbench.addons.junit5.extensions.unittest.VaadinTest;
-import junit.com.vaadin.testbench.tests.uitest.GenericTestPageObject;
+import junit.com.vaadin.testbench.tests.ui.GenericTestPageObject;
 import org.junit.jupiter.api.Assertions;
 
-import static com.vaadin.testbench.tests.uitest.ElementQueryView.ROUTE;
+import static com.vaadin.testbench.tests.ui.element.ElementQueryView.ROUTE;
 
 @VaadinTest
 class ExecuteJavaScriptIT {
