+++ vaadin-testbench-it/vaadin-testbench-it-test/src/test/java/junit/com/vaadin/testbench/tests/ui/element/PolymerTemplateViewElement.java
@@ -1,4 +1,4 @@
-package junit.com.vaadin.testbench.tests.uitest.elements;
+package junit.com.vaadin.testbench.tests.ui.element;
 
 import com.vaadin.testbench.TestBenchElement;
 import com.vaadin.testbench.annotations.Element;
